go to def for record names in is_record/2 or /3
Summary:
The second argument to `erlang:is_record/2` and `erlang:is_record/3` is the name of a record.
Make sure go to def reacts to this, including checking the size for `erlang:is_record/3`.

This also fixes some unused include diagnostics that did not realise we were referring to records defined in them.

Reviewed By: robertoaloi

Differential Revision: D52120347

fbshipit-source-id: 11526329696fbc8420898f0fa627adf0cf3a3a30