Signal diagnostics found for warnings as well as errors
Summary:
We use `elp lint` as part of our regular ELP CI, and waserver has many warnings on it. But we want to know if we introduce or find any error diagnostics.  To achieve this, we only give a failing exit status if errors are found, not if only warnings are.

BUT, when we use this for WASERVER CI, it means warnings do not show up.

We use JSON output for the latter, so give the failing status for errors and warnings in that case.

Reviewed By: VLanvin

Differential Revision: D52170371

fbshipit-source-id: 495643836fe05ea4f476a281ce69a9d38e33f710