Add diagnostic for simplifying expressions using algebraic identities
Summary:
# Context
When writing codemods, one may want for instance to remove elements from a list (e.g. props that are no longer needed) and the list may become empty. That means that one can go from:

```
[{foo, Bar}] ++ MoreProps
```

to

```
[] ++ MoreProps
```

One would like further simplify this to just `MoreProps`.

# This diff
So we introduce a new diagnostic that will be displayed when an expression contains an expression that can be simplified according to algebraic identities. For now, we handle only some list and arithmetic identities, but this can be easily extended in the future to account for other identities, known functions calls, etc.

Reviewed By: alanz

Differential Revision: D52261493

fbshipit-source-id: 38db044fffbd4b73f89f063bb67e11012b0cc80d