Move extraction arg from first clause to FunctionDef
Summary:
Move the logic to extract arguments from the first clause of a function to the `FunctionDef`, to make it more discoverable and to avoid spreading logic around.

Also avoid an unwrap in the arg names extraction, making the returned type optional. When using the functionality in the context of an `add_impl`, adds an expectation since the assumption should always hold.

Reviewed By: alanz

Differential Revision: D52796672

fbshipit-source-id: 920ab740cdb0828387b17bfbc4dfba786f9b2094