Consistently use DiagnosticsConfig::default method
Summary:
A `default()` method in combination with setter, it becomes less verbose to use a default configuration (as opposed to a `new()` one, where all parameters must be specified. At the same time, it makes more obvious the intention of the caller when modifications to the config need to be made.

This is to simplify the merge between `DiagnosticsConfig` and `ErlangServiceDiagnosticsConfig` later in the stack.

Reviewed By: alanz

Differential Revision: D52953186

fbshipit-source-id: b0ba077bacdc677e0a1896580244f0a293bae539