1/n migrating project app data into Project.apps
Summary:
In ELP, the `Project` data structure is used as the source of truth for operating on a given set of files.
This is generated from test fixtures, and from the various project types, buck, rebar, static file.

It has a field `Project.project_build_data` which tracks the origin.  All of these also include some variation of `apps: Vec<ProjectAppData>`.  The `ProjectAppData` describes a particular Erlang application, and has an `AppType` flag in it indicating whether it is part of OTP, a dependency or an app.

But once we start processing the `Project`, we no longer care about the origin, only the `ProjectAppData`.

This series of diffs migrates it into a top level field `Project.project_apps`.

Reviewed By: michalmuskala

Differential Revision: D53130628

fbshipit-source-id: bc75f0c5258536f44f45488ef8b516d9f1e59ef7