fix: show edoc assist on top level functions of modules with edoc comment
Summary:
## What

Look up current function's edoc using `file_edoc_comments` (which was not available when this code was originally written).

## Why

Previously, when the Erlang module had module-level documentation, the lookup would return module-level documentation as edoc for the first function.

As a result, edoc assist was always broken for the first function of every Erlang module with module-level edoc.

Reviewed By: alanz

Differential Revision: D53181834

fbshipit-source-id: f00ea9f04d1fde757b5c8b04669f46c1408af906