unused header detects record name in macro RHS
Summary:
We sometimes define a record name as a macro. For reasons.
So if we find an atom in the RHS of a macro, check if it is referencing a record name, and mark that as a reference.
This allows us to more robustly detect if an include file with record definitions is being used.

Reviewed By: TD5

Differential Revision: D53185387

fbshipit-source-id: 6fd91ebd96ba61f4ef4673579ea739dbf2336a34