1/n Declarative Tests: Add meta info to tests for enabled diagnostic sources
Summary:
This is the first of a series of diffs that will result in us being able to declaratively configure which code intelligence sources we have available in our tests.

This first one introduces `DiagnosticsEnabled` and populates it with flags for the erlang_service and common test, populated from declarative test metadata.

They do not have any effect yet.

Reviewed By: robertoaloi

Differential Revision: D53309970

fbshipit-source-id: 2d101a35bb03362ce8ec9f770c13f20c30ee4db8