BE: 1/2 put a Semantic field into InFunctionClauseBody
Summary:
`InFunctionClauseBody` presents the primary API for working with the HIR ast on a function clause.
Many of the API methods pass in a parameter of either `Semantic` or `MinDefDatabase`.  `Semantic` contains one of these databases.

Since we create an `InFunctionClauseBody` in the context of `Semantic`, store a copy of in it for later use.

This is the first of two diffs, which just stores it. The next will simplify the API calls to use it.

Reviewed By: ir-regular

Differential Revision: D53307828

fbshipit-source-id: 9045e7a9fed325a8c1249c0228b3f221c57bd607