5/n Declarative Tests: action '//- common_test' for test fixture
Summary:
Update the tests needing to run the common test helper via the erlang service to do so via declarative metadata.

This is the first rough cut at how it is done in the test infrastructure, but once the tests are annotated the drivers can change without disturbing the tests further.

Reviewed By: robertoaloi

Differential Revision: D53426828

fbshipit-source-id: 54c2d0957f93c659bcd47eea1a9fe163effb1c91