4/n Declarative Tests: fixture::position returns DiagnosticsEnabled
Summary:
We parse a declarative test fixture and return a fixture and changes to be applied to the database, used both for `TestDB` and `RootDatabase`.
If we are running against a `RootDatabase` (i.e. for `ide` crate tests), we can also start up the erlang_service, eqwalizer, and all other complex services.
So return a `DiagnosticsEnabled` extracted from the test fixture metadata, to do this.

This diff only returns the data, and updates all call sites to match.

Reviewed By: robertoaloi

Differential Revision: D53426489

fbshipit-source-id: 7d095272ad16a703fc8e5f725ed0dec3a9caa0da