7/n Declarative Tests: move calculation of eqwalizer per-file diagnostics to Analysis
Summary:
The `elp` crate has a `Snapshot` struct that is used in the LSP server, and has utility functions in it.
Move the one to calculate Eqwalizer diagnostics out of it and into `Analysis`, which is the equivalent concept in the `ide` crate.

This allows us to use it for declarative `ide` tests later in the stack.

Reviewed By: TD5

Differential Revision: D53427904

fbshipit-source-id: ac735a017d513b4fbc8869073eff9afca2b31e74