15/n Declarative Tests: Test Fixture: dump full fidelity to temp dir
Summary:
If a test fixture metadata requires files on disk, dump the project to disk in a temporary directory, with a build_info.json file describing it, then load that for use in the tests.

We do not make use of anything requiring on-disk access yet, this comes later in the stack.

Reviewed By: perehonchuk

Differential Revision: D53477256

fbshipit-source-id: 89cb09845e1793180069a432737b8554ebcb3959