11/n Declarative Tests: prevent build info being deleted prematurely in tests
Summary:
When we load a project we generate a temporary directory holding the build info needed by eqwazizer.
This has an on drop method which cleans it up when it goes out of scope.
Ensure it stays in scope for the duration of the test by storing the test fixture projects into the `DiagnosticsEnabled` structure, which is used to start up the processes that need the build info, so tracks the required lifetime.

This only takes effect later in the stack.

Reviewed By: perehonchuk

Differential Revision: D53429732

fbshipit-source-id: 301a0fc3997bdf3f93a42d6608668eaf8fe632db