20/n Declarative Tests: eqwalizer tests
Summary:
Confirm that our infrastructure is now sufficient to enable eqwalizer for a test.

Note that we check for a string representation of the Eqwalizer diagnostic code, as a simplification in setting up the test fixture.

In future we can have other message translators providing appropriate information for matching in the tests.

Reviewed By: perehonchuk

Differential Revision: D53479131

fbshipit-source-id: bc3920284d5ac89d1d2a2b6f3d608836d06ff7d2