9/n Declarative Tests: make a copy of RA testdir.rs as temp_dir.rs
Summary:
We do this because we need fine-grain control of when to drop the TempDir, which the stdlib version does not give us.
We need to be able to create it, assign it to a structure, and only have it cleaned up when that structure goes out of scope.

It is also convenient when debugging tests to be able to easily keep the temp dir around for inspection once the test has run.

This initial diff just imports it, we will use it later in the stack.

Reviewed By: perehonchuk

Differential Revision: D53429012

fbshipit-source-id: e008d85153257d3021d7cb5f2febc65d20ecfa08