13/n Declarative Tests: Introduce a test Builder
Summary:
When we need to save a fixture to disk, the paths in vfs must match the actual location onnn the disk.
The `Builder::new` is called with the `DiagnosticsEnabled` being used, and will only create a temporary directory if needed.

It provides a method to retrieve an absolute path, either to the location on disk or as given in the test fixture.

We use it further up the stack.

Reviewed By: perehonchuk

Differential Revision: D53430463

fbshipit-source-id: 68a2da27738315b8fdba8225f1b1c0156958baf9