17/n Declarative Tests: drop-proof annotation test fixtures
Summary:
The tests for annotations wrap the normal test fixture setup, then run the tests.
For tests requiring an on-disk fixture, this means it is dropped and deleted before the test runs.
So return a `DiagnosticsEnabled` structure that can be kept in scope until the tests finish, to ensure the fixture is deleted at the right time.

Does nothing but plumb through the new parameter to do this.

Reviewed By: perehonchuk

Differential Revision: D53478440

fbshipit-source-id: 4c2ef328b68f099f22fe534120f37e4ddbfbc29e