12/n Declarative Tests: Do not populate ebin dir in test fixture
Summary:
When we load a project via a JSON config file, we check that all the listed directories exist.
We had been setting `ProjectAppData.ebin` to directory named "ebin", but none of our tests use it.
So the simplest way to prevent project loading errors is to not populate the field.
This takes effect higher up the stack. Getting close now.

Reviewed By: perehonchuk

Differential Revision: D53430023

fbshipit-source-id: 279b0059c48cb7beda26dfef9accfa91b6848e67