Remove nested syntax errors
Summary:
The tree-sitter parser can give us a parse tree with error nodes in it, but with valid parse information contained within it for the parts it was able to parse. But sometimes the interior nodes have additional errors.

At the moment we report all of them, which means we end up with large error spans from a syntax error.

We now rather filter out any higher lever syntax error that contains any lower level ones.

Reviewed By: michalmuskala

Differential Revision: D53573429

fbshipit-source-id: dabd9ceedaff28fe6b9a40d4f77faa4b3c08b879