Solve panic for build_info command in presence of complex macros
Summary:
According to the [Erlang documentation for `erl_opts`](https://www.erlang.org/doc/man/compile.html):

```
  {d,Macro}
  {d,Macro,Value}

Defines a macro Macro to have the value Value. Macro is of type atom, and Value can be any term. The default Value is true.
```

The handling of user-defined macros via the `build_info` mechanism is currently very basic in ELP and only the first format is supported.

While this change does not aim at providing full support for custom macros, we:

* Change the data structure from a Vector to a HashMap (so we can store the value)
* Default the value to the atom `true`
* Only support atoms for the time being

Reviewed By: alanz

Differential Revision: D53918409

fbshipit-source-id: 650c900d6c254c22b99c3b653f121b9c5d1b0bed