Add TypedSemantic to DefDatabase
Summary:
`DefDatabase` is the primary salsa-backed database used in the HIR crate.

In particular,  `struct Semantic` contains one, and it is used to provide the main API for programming against HIR.

Note that when we construct a `Semantic` instance in the `IDE` crate, it uses a `RootDatabase`, which contains a `DefDatabase` but also the information needed to run Eqwalizer.

By adding `TypedSemantic` to it, we make the Eqwalizer types available too, but only when used from the `IDE` crate.

`DefDatabase` is also included in the `TestDB`, which is used as the basis for fast declarative tests in HIR.  We add an implemetation for the `TypedSemantic` trait which simply panics with a warning that the calls are not supported in HIR only.

In future we may provide mock implementations to allow this via declarative tests.

Reviewed By: michalmuskala, VLanvin

Differential Revision: D53916646

fbshipit-source-id: 3a155b51f99bb9b7b4820f62d44a7193aebd89d2