Introduce elp_types_db crate
Summary:
This stack is about updating the `Semantic` API to be able to programme with inputs over and above ones from HIR. These, like Eqwalizer, require external processes and so cannot be used in the HIR crate directly, but are available in the IDE one, which depends on HIR.

So we take a strategy of introducing a new crate, used in HIR and IDE, which just provides the types to be used in the API.  The provided API using these types returns `Option<_>` values, so will fail in HIR, but can be populated in IDE.

It also allows us in future to provide implementations in `TestDB`, which can give us high-speed declarative tests of IDE features using a mock implementation of the data sources.

This Diff simply does the housekeeping around introducing the crate.

Reviewed By: michalmuskala, VLanvin

Differential Revision: D53851948

fbshipit-source-id: 75f2f2739ac37b7edad530eedce04ed811db2667