Introduce meck linter
Summary:
This change introduces a `meck` linter, which can be used (with the limits imposed by a static analysis type of check) for common `meck` pitfalls.

The first check implemented is the absence of the `no_link` option in calls to `meck:new/1,2` inside a `init_per_suite/1` or `init_per_group/2`. More checks will be added later. Error code `W0022` is introduced for the check, accompanied by an Erlang Error Index entry.

Reviewed By: alanz

Differential Revision: D53940743

fbshipit-source-id: b922e40d300d40f6eb59acc28f55413c76e837f4