unwinding for edoc
Summary:
**Problem**: erlang service blocks main loop by holding the read lock on salsa database

**Solution**: instead of blocking `recv()` use `recv_timeout()` in a loop and after each timeout check if request was cancelled with the closure which refers to `db.unwind_if_cancelled`. This way we will block not more than 100ms and not have direct dependency to `base_db` in `erlang_service` crate.
In case of `unwind` receiver will be just dropped and `crossbeam` would return `err` on sending the response.
Note that this solution will unblock the main loop, but `erlang service` will still perform the work. As a follow-up we can introduce canceling on the erlang service side.

This diff adds unwinding for **edoc**. Other requests will be covered in next diffs

Reviewed By: michalmuskala

Differential Revision: D54193392

fbshipit-source-id: 0f2485a0024194dce4c2242dc0e0de1db7f6ae20