Introduce FileKind::TestModule
Summary:
We have quite a few places where we check for test suites by looking at the `_SUITE` suffix of a file. Instead, introduce a new `FileKind` which allows us to distinguish between `SrcModule` and `TestModule`.

Notice that this is orthogonal to the `extra` file kind used by `is_test_suite_or_test_helper` (out of scope for this change).

Reviewed By: michalmuskala

Differential Revision: D54255637

fbshipit-source-id: f884a73fd18fc6674f8e95d7161921809c3a91ce