Add origin to Body
Summary:
Store the origin of a `Body` in it, as the parameter to be used to the appropriate XXX_body() call that produced it.
This allows us to retrieve the matching `BodySourceMap` if needed.
And unfortunately it is needed to tie up HIR ast items with Eqwalizer types.
Hopefully in the future this need will go away.

Reviewed By: michalmuskala

Differential Revision: D53710616

fbshipit-source-id: 366eff9ca35ca7ce834c9c96c400761a8db5c725