Introduce TypedMFA for FunctionMatcher
Summary:
This allows `find_call_in_function` to match on the types of the arguments in a given call.
Note that this relies on Eqwalizer to have determined the types for all arguments, otherwise the match fails.

This is based on a diff by robertoaloi

Note that this should be regarded as an experimental API. It will only match if eqwalizer does have types for each argument in the call, and will silently fail otherwise, with no indication as to whether it was a type mismatch or simply not having types available.

It is probably best to use it for assists only initially, which are opportunistic, rather than diagnostics.

Reviewed By: michalmuskala

Differential Revision: D53711416

fbshipit-source-id: e6aeb48c2b1ea243f843f5ba140d438fb5e0df18