Semantic::expr_type
Summary:
The first extension to the `Semantic` API from having `TypedSemantic`.

Provide a convenience function for an `ExprId` that looks up its range via `BodySourceMap`, and calls  `eqwalizer_type_at_position` on it.

At some future date we may be able to dispense with the range lookup, but the API should then be able to stay the same.

Note: It is only possible to retrieve types for variables in function bodies, and the subjects of `case` statements. But we provide a generic API as we need to be able to use it from within our `fold` infrastructure.  The API returns `Option<eqwalizer::Type>`, so the invalid cases can be handled cleanly.

Reviewed By: michalmuskala

Differential Revision: D53917688

fbshipit-source-id: 0f7418fe6ed508a771f537c14993174084f0d158