BE: Add ELP config for erlang_service
Summary:
Now that we can use an `.elp.toml` file to point to a `build_info.json`, we may as well do it.

FYI, the build info file was generated by temporarily adding the build info plugin, running `elp build-info --json --to build_info.json` and then removing it again.

Reviewed By: perehonchuk

Differential Revision: D54363342

fbshipit-source-id: b1696d9a5cc7165ba5715848f3cc6cfcf09e7026