BE: Track included header files for erlang service module_ast()
Summary:
The abstract forms returned by the Erlang Service when parsing a module include `file` `attributes` for each file included while it is processed.
This diff extracts the list of files in the Erlang Service, and registers a dummy read against salsa so salsa can track the dependency tree for it when loading the ast in the rust side.

This means the `include_file_revision` hack to conservatively invalidate *all* files if a `.hrl` file changes can be removed. This is done in the next diff.

Reviewed By: michalmuskala

Differential Revision: D54357782

fbshipit-source-id: 168c76c3128bcb7390e5e2df5271532dd5a5433f