Introduce parents in fold callback context
Summary:
The callback in a fold provides a context, which has the specific HIR item being visited.
But it is impossible to know anything about its parents.
Fix this by keeping track of the parents up to the root of the HIR AST in a body, and passing this list in `AnyCallBackCtx`.

Reviewed By: robertoaloi

Differential Revision: D54262650

fbshipit-source-id: ea13cf1ad586a79be3e22b76721dcf463efcbc43