rework combining diagnostics: split labeled into undefineds and syntax errors
Summary:
We can have a situation where Erlang Service reports a diagnostic for a FunDecl, but ELP does not.
We need to add these to related info and deal with undefined function diagnostics arising from them too.
As a first step, split the labeled diagnostics field in two, so that in a following diff the Erlang Service can return both syntax errors by function and diagnostics about undefined functions.

This diff just does the split, no operational change.

Reviewed By: michalmuskala

Differential Revision: D54492993

fbshipit-source-id: 02e3e6c902c383a98a15e248c4267855c614f128