Flakiness: Introduce a lock to serialise check_type test
Summary:
Tests using this function are occasionally failing in CI.
I suspect it is because of running more than one erlang_service instance at the same time.
Protect the function with a lock to check the hypothesis.
It may not help, because other tests using the erlang_service have a static connection.
Dealing with that will be step 2 if this does not help.

Differential Revision: D54631248

fbshipit-source-id: 842f601c815524bfa84957f959a7ac2cc36452bf