Put mutex around erlang service for tests
Summary:
We are seeing flaky test results on continuous runs related to tests accessing type info.
At the moment, we run a single instance of the erlang service, which is not re-entrant.
Guard access to it in tests with a global mutex.  This follows the same logic as for temporary directories used in tests, we return a locked guard, which is released when the reference is dropped as the test ends.

Reviewed By: robertoaloi

Differential Revision: D55063978

fbshipit-source-id: 18bfd3f9aed1042ee983bca5bfecd17f9e3a377c