Type Assists 3/n: initial type based assist, for mismatched atom
Summary:
This is the simplest possible type-based assist, to confirm that the entire process works.
When there is a spec saying a function returns an atom, if an incorrect atom is returned offer to change the returned atom.

We will extend this later to change the spec, but that requires a bit more code, this is intended to be the simplest possible for now.

Reviewed By: robertoaloi

Differential Revision: D54956646

fbshipit-source-id: e60c56d8b128147eb92541bc3668c2532b2126c1