Type Assists 1/n: Add eqwalizer assists to code_action request handler
Summary:
This is the start of a stack to make use of the recent ELP internal refactor and new structured type errors from Eqwalizer to provide assists based on Eqwalizer diagnostics.

The first step is to ensure that Eqwalizer diagnostics that have associated fixes get plumbed through to the LSP client when a code action request is received.

This is done by adding the current LSP server `DiagnosticCollection` to the snapshot, and processing the Eqwalizer diagnostics in it when looking for assists.  This approach allows us to add assists to the Erlang Service OTP diagnostics, edoc diagnostics, and so on too, in future.

Note that the current architecture does not update the Erlang Service or Eqwalizer diagnostics on file change, only on save. So the assists will be based on the  most recent diagnostics seen in the `DiagnosticsCollection`. Hopefully we can improve this in future.

Reviewed By: robertoaloi

Differential Revision: D54956261

fbshipit-source-id: caa57b541442e112350d1657cbfe2afb388415ef