BE: make contains_annotation regex more lenient
Summary:
Allow multiple spaces between the last position marker and the start of the content.
The `contains_annotation` test is used to strip annotations from a file.
Before this change, an annotation with extra spaces would be recognised and processed by the annotations parser, but not stripped by a call to `contains_annotation`

Reviewed By: robertoaloi

Differential Revision: D55071427

fbshipit-source-id: 7256fd55d49347c9663aa90e5a9a53f7eceb0386