Flakiness prevention by unique modules
Summary:
We have a problem with flaky tests which use the eqwalizer types API, which in turn uses the erlang service.

We only have a single instance of the Erlang service during the tests, to avoid wasted time starting and stopping an Erlang VM.

But within the Erlang service modules are loaded by module name only, and if more than one test uses the same module name we can end up with incorrect information returned.

This diff protects against that by keeping a static registry of all module names we spill out to disk when setting up a test fixture for use in the Erlang Service.  If we attempt to use one already seen in the run, panic with the offending module as well as a list of all modules seen so far, to help in choosing a new unique name.

Reviewed By: robertoaloi, perehonchuk

Differential Revision: D55368720

fbshipit-source-id: ccffc98d09708b91554f4c08f31802cf825ccdf3