Add undefined function in export diagnostic as related info if syntax error in the function
Summary:
This is an extension of the existing usage for syntax errors reported natively in ELP, when we have an erlang service syntax error and not a native one.

Also make the allocation of a syntax error to a function more robust, by looking for an immediately-preceding FunDecl if our syntax error is in a top-level error node.

Reviewed By: michalmuskala

Differential Revision: D54493175

fbshipit-source-id: 0a96651afa31eedf8f3ca60999f622893b6a8fbe