Bump eetf
Summary:
This introduces a new explicit type for the `ByteList` optimisation which makes some conversions more efficient and requires less allocations.
It also changes the representation of `eetf::Map` to a hashmap (rather than a vec of pairs). We take advantage of that by making the conversions from build_info more efficient avoiding many re-allocations.

Reviewed By: TD5

Differential Revision: D55418250

fbshipit-source-id: ac85b42bf8243992fa67d53ae94c8c173223f653