Preliminary support for project-wide Eqwalizer diagnostics
Summary:
This is an initial, sequential, implementation of project-wide eqwalizer diagnostics. The implementation will be refined and parallelized later. Behind a feature flag, disabled by default.

Project-wide diagnostics are requested contextually with the regular eqwalizer diagnostics for the current module. The duality is required to keep per-module diagnostics fast.

One difference with regular diagnostics is that project-wide diagnostics are not requested on open.

Opened files are excluded from the project-wide diagnostics to avoid duplicates.

The extra logic for handling the diagnostics collection is related to the need of clearing existing diagnostics for files who are not in the result.

Reviewed By: VLanvin, alanz

Differential Revision: D55316244

fbshipit-source-id: 77b9aa99ea0d37cde80e60de82b21a3d315475ab