Diagnostics: Add helper to declaratively generate a diagnostic if a function is used
Summary:
This brings in a `DiagnosticTemplate` for a (static) diagnostic, giving message, code, severity, whether it should include a quick fix, and a structure to capture function matches.

This allows us to generate a given diagnostic for a function being used by just calling a function with the appropriate data structure.  It is made use of in the next diff.

Reviewed By: robertoaloi

Differential Revision: D55691840

fbshipit-source-id: 1666e06972051750f475f9702f005264598e3e1a