Introduce a diagnostic descriptor, to manage when they should run
Summary:
Move toward a diagnostic source being captured in a descriptor, which provides the conditions for when it is applicable and a pointer to a function to generate it.

This allows us to filter out inapplicable ones in the main diagnostic loop, and can provide more sophisticated management in future, dealing with whether individual diagnostic codes or categories are enabled.

This is the initial proof of concept, being applied to a single diagnostic.

Reviewed By: robertoaloi

Differential Revision: D55699081

fbshipit-source-id: 4dbbe4b97aecac5a0b38f24306942cd905bbbfe6