Go To Type Definition
Summary:
This change adds support for the [textDocument/typeDefinition](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_typeDefinition ) LSP request. The request is sent from the client to the server to resolve the type definition location of a symbol at a given text document position.

In simpler words, it enables the user to use the "go to type definition" menu (or to use a shortcut) to jump from a variable to the place where the type of that variable is defined. In the case of union types, the user previews the different possible locations and jumps to one of them.

Notes:

* Eqwalizer returns an approximate range for the type references, so we need to adjust it to be able to reuse the Classify infrastructure
* Eqwalizer currently does not return references for OTP functions. If that gets solved, we'd also get navigation for OTP types

Reviewed By: alanz

Differential Revision: D55736888

fbshipit-source-id: a1f8909e069228e90f57f82e1998ac2c3b89342f