add Diagnostic.with_fixme_fix
Summary:
We have an `elp:ignore` comment that suppresses display of the named diagnostic, indicating that it is spurious, the code is intended to be that way.

Introduce a `elp:fixme` comment which also suppresses display of the named diagnostic, but indicates the problem should be fixed in future.  This can be used for codebase quality management.

It is unused at the moment, will be used further up the stack.

Reviewed By: robertoaloi

Differential Revision: D55793040

fbshipit-source-id: 2821e2ece9125afd468bbffd927267f777fe5c0c