Remove test module uniqueness checker
Summary:
Since test module uniquess is not after all a cause of test flakiness, we no longer need to burden ourselves with the discipline of having unique module names across all test fixtures.
This diff does not back out the initial name changes when it was introduced, they can stay.

Reviewed By: robertoaloi

Differential Revision: D55964757

fbshipit-source-id: 4f2f2d2988319b4e20923dea09bf464ed2796403