Only create the eqwalizer exe once
Summary:
We have a complex process of identifying and setting up the precise eqwalizer to use for a given run.
This is implemented in the `Eqwalizer::default()` method, and historically was safe.

We are now running tests that use the erlang service, and in the process create a fresh RootDatabase (calling `Eqwalizer::default()`). Each test can run in a separate thread, so we can have multiple concurrent initializations.  Under normal usage the Eqwalize executable is bundled inside ELP, and must be dumped to disk for use.  The mechanism to decide on the directory has a race conditition, where if it is called from two different threads at the same time each can choose the same file name.
This means one can write and start running, but when the second tries to launch it gets a "text file busy" operating system error.

We avoid this, and also speed up the tests, by doing the initialization once, storing the result in a lazy static variable.

Reviewed By: robertoaloi

Differential Revision: D55964394

fbshipit-source-id: 61805047323c2c863ecf3bed272c0935b30d14bc