doc declarations for macro
Summary:
Currently glean supports showing comments from declarations on xref hover.
The problem is for the macro it works poorly, since we need data from the calling part.
This diff makes the same trick with macro like previous with vars: create a definition right on the xref location and add relevant comment there. As arity macro xref location is used. This way we can have ODS links, macro expansion and macro definition on hover. Drawbacks is that we miss go-to-def and find references for macros

Reviewed By: alanz

Differential Revision: D55916496

fbshipit-source-id: 29fa4d399516792172d592faf8dcd4699bb3265a