add_fixme assist
Summary:
We have two ways of providing code actions in ELP. The first is to add it directly to the Diagnostic.
The second is to use the full code action request machinery, and match it to the original diagnostic code.
The latter is more complex, but allows us to use the `resolve code action` request and middleware to request user input for the assist.

This is what we do here when adding a fixme comment, where the additional information can be requested from the user in clients that support this mechanism.

Note: later in this diff stack the user input is extended to include requesting a task id

Reviewed By: perehonchuk

Differential Revision: D55912943

fbshipit-source-id: b12c77ded3d5264b01445e3eced85b31f49ede7f