elp lint: bringing in code action fixes
Summary:
Fixes can originate from two places: they are either calculated as part of the original diagnostic, and so included in them, or generated via a request for a code action.
Up to now `elp lint` has only been able to apply the first kind.

This diff extends it to apply assists of the second kind too, but only if they are specifically generated in response to the diagnostic being considered at the time.

Reviewed By: perehonchuk

Differential Revision: D55961937

fbshipit-source-id: ac17637757faa5a3b046c791cb3e1c0986f42df7