BE: Enable eqwalizer types when using `elp lint`
Summary:
Bring the server and `elp lint` eqwalizer environments into alignment.

Recent changes in the Eqwalizer internals have erased the practical differences between the `Eqwalizer::Mode` exposed in ELP in terms of operations, except the vital difference that type information is currently only returned in `Server` mode. So use this for `elp lint`, to be able to query types in lints.

Also start erlang service for OTP, since we do this in server mode.

Reviewed By: robertoaloi

Differential Revision: D56134199

fbshipit-source-id: f262c409b29aafa89f48740086093a11c9f09141