specify erl path from cli
Summary:
in buck env everything should be hermetic, so we can't rely on erl in the path.
this diff adds an option to specify path to `erl` binary for any command.
If the option is not specified `elp` will use `erl` from `$PATH`

this option will be used in D55965578

Reviewed By: alanz

Differential Revision: D56197691

fbshipit-source-id: 51f735d25e5006c558b7abdb80a0a53b7a0ea256