sort out panic for get_file_contents on a created file
Summary:
We are seeing panics where the panic context indicates the file change operation is `Create`.
Having exhaustively studied the elp server.rs and vfs code, it appears that the file contents have to be valid for a `Create`.

And the only way we can have a `FileId` with contents of `None` (which is what triggers the panic) is if there has been a delete operation on the file.

In the server we do not process changes as they happen, we let the vfs loader process batch them up for us, then process the batch.

So it seems that the only way the problem can happen is
- a file is created
- then it is deleted
- and we process both changes in the same batch
In this case we panic on the first, because the second change is already in vfs, from the loader thread.

To mitigate this, lock vfs while processing changes, and call `vfs.file_id()` to check that the given file_id has not been deleted in the locked vfs before processing it.

Reviewed By: michalmuskala, robertoaloi

Differential Revision: D56352372

fbshipit-source-id: 018be8dae4022d422bc81ed7e15a2dbd96f45cfd