Avoid project-wide diagnostics disappear upon closing
Summary:
When closing a file, we were resetting all diagnostics for that file. Now that we have project-wide diagnostics for eqwalizer, this behaviour is incorrect, since it causes project-wide issue for the same file to disappear upon file closure.

Exclude project-wide diagnostics when resetting diagnostics.

Notice that, to keep eqwalizer diagnostics for the local module fast, we keep them separate. So we just move them around, from local to project-wide.

Also, do not exclude opened files from project diagnostics, or we can end up with missing diagnostics. Duplicates are handled at collection level, anyway.

T186514003 is the task to reconsider local vs global eqwalizer diagnostics, but for the time being this should solve the current bug.

Reviewed By: alanz

Differential Revision: D56253723

fbshipit-source-id: e7998091166b4e3af6f18ced30546bf1ece9ddc6