BE: use the same processing path converting from vfs to salsa
Summary:
We have a number of places that take the contents of a file from vfs and store them in the salsa database.
This processing includes converting from `[u8]` to a string in a safe way, and normalising line endings.
Put the processing in one place, and use it everywhere.

Note that the line_endings information is only used when generating output again, as in assists with edits. So it is ignored in the shell and CLI cases.

A later diff will make use of it for `elp lint`.

Reviewed By: robertoaloi

Differential Revision: D56474440

fbshipit-source-id: a7e82dd6a64e94fa07d840fcf7baebfb1f07ae29