Update erlang_service to process CRLF the same as ELP
Summary:
In ELP we normalise all internal representations of file to use LF only, never CRLF, and convert when loading.
The erlang_service does not do this, so for files using CRLF the diagnostic locations do not line up.
Update the erlang_service to not count CR when computing positions, so it lines up with ELP

Differential Revision: D56514998

fbshipit-source-id: 101584d59d07bec28976d1451e98436ab2c9c986