Capture LineEnding in LoadResult
Summary:
This is a follow on from the prior diff.
We store the line ending map in `LoadResult`, and populate it when loading a file.
Currently it is unused, the only place the line ending map is used is in converting assists sent to the LSP client.

Note: it is not needed for `elp lint` assists, as we work on the internal normalised representation.
Perhaps it should take cognisance of this when writing out changed files, in a follow up diff.

Differential Revision: D56514878

fbshipit-source-id: e96cee5c7bf9e1343e8b7540dccc7ab5aff678bf